+++ ingenialink/dictionary.py
@@ -136,6 +136,18 @@ class DictionaryError:
         return iter((id_hex_string, self.affected_module, self.error_type, self.description))
 
 
+@dataclass
+class DictionaryDescriptor:
+    firmware_version: Optional[str] = None
+    """Firmware version declared in the dictionary."""
+    product_code: Optional[int] = None
+    """Product code declared in the dictionary."""
+    part_number: Optional[str] = None
+    """Part number declared in the dictionary."""
+    revision_number: Optional[int] = None
+    """Revision number declared in the dictionary."""
+
+
 class Dictionary(ABC):
     """Ingenia dictionary Abstract Base Class.
 
@@ -226,6 +238,20 @@ class Dictionary(ABC):
             self.read_dictionary()
         except KeyError as e:
             raise ILDictionaryParseError("The dictionary is not well-formed.") from e
+
+    @classmethod
+    @abstractmethod
+    def get_description(cls, dictionary_path: str, interface: Interface) -> DictionaryDescriptor:
+        """Quick function to get target dictionary description
+
+        Args:
+            dictionary_path: target dictionary path
+            interface: device interface
+
+        Returns:
+            Target dictionary description
+        """
+        pass
 
     def __add__(self, other_dict: "Dictionary") -> "Dictionary":
         """Merge two dictionary instances.
@@ -507,6 +533,28 @@ class DictionaryV3(Dictionary):
     __PDO_ENTRY_SIZE_ATTR = "size"
     __PDO_ENTRY_SUBNODE_ATTR = "subnode"
 
+    @classmethod
+    def get_description(cls, dictionary_path: str, interface: Interface) -> DictionaryDescriptor:
+        try:
+            with open(dictionary_path, "r", encoding="utf-8") as xdf_file:
+                tree = ET.parse(xdf_file)
+        except FileNotFoundError as e:
+            raise FileNotFoundError(
+                f"There is not any xdf file in the path: {dictionary_path}"
+            ) from e
+        root = tree.getroot()
+        device_path = (
+            f"{cls.__BODY_ELEMENT}/{cls.__DEVICES_ELEMENT}/{cls.__DEVICE_ELEMENT[interface]}"
+        )
+        device = root.find(device_path)
+        if device is None:
+            raise ILDictionaryParseError("Dictionary cannot be used for the chosen communication")
+        firmware_version = device.attrib[cls.__DEVICE_FW_VERSION_ATTR]
+        product_code = int(device.attrib[cls.__DEVICE_PRODUCT_CODE_ATTR])
+        part_number = device.attrib[cls.__DEVICE_PART_NUMBER_ATTR]
+        revision_number = int(device.attrib[cls.DEVICE_REVISION_NUMBER_ATTR])
+        return DictionaryDescriptor(firmware_version, product_code, part_number, revision_number)
+
     @staticmethod
     def __find_and_check(root: ET.Element, path: str) -> ET.Element:
         """Return the path element in the target root element if exists, else, raises an exception.
@@ -531,8 +579,8 @@ class DictionaryV3(Dictionary):
         try:
             with open(self.path, "r", encoding="utf-8") as xdf_file:
                 tree = ET.parse(xdf_file)
-        except FileNotFoundError:
-            raise FileNotFoundError(f"There is not any xdf file in the path: {self.path}")
+        except FileNotFoundError as e:
+            raise FileNotFoundError(f"There is not any xdf file in the path: {self.path}") from e
         root = tree.getroot()
         drive_image_element = self.__find_and_check(root, self.__DRIVE_IMAGE_ELEMENT)
         self.__read_drive_image(drive_image_element)
@@ -617,7 +665,7 @@ class DictionaryV3(Dictionary):
         else:
             device_element = root.find(self.__DEVICE_ELEMENT[self.interface])
         if device_element is None:
-            raise ILDictionaryParseError("Dictionary can not be used for the chose communication")
+            raise ILDictionaryParseError("Dictionary cannot be used for the chosen communication")
         self.__read_device_attributes(device_element)
         if self.interface == Interface.ETH:
             self.__read_device_eth(device_element)
@@ -1054,12 +1102,51 @@ class DictionaryV2(Dictionary):
         },
     }
 
+    _INTERFACE_STR = {
+        Interface.CAN: "CAN",
+        Interface.ECAT: "ETH",
+        Interface.EoE: "ETH",
+        Interface.ETH: "ETH",
+    }
+
+    @classmethod
+    def get_description(cls, dictionary_path: str, interface: Interface) -> DictionaryDescriptor:
+        try:
+            with open(dictionary_path, "r", encoding="utf-8") as xdf_file:
+                tree = ET.parse(xdf_file)
+        except FileNotFoundError as e:
+            raise FileNotFoundError(
+                f"There is not any xdf file in the path: {dictionary_path}"
+            ) from e
+        root = tree.getroot()
+        device = root.find(cls.__DICT_ROOT_DEVICE)
+        if device is None:
+            raise ILDictionaryParseError(
+                f"Could not load the dictionary {dictionary_path}. Device information is missing"
+            )
+        dict_interface = device.attrib.get("Interface")
+        if cls._INTERFACE_STR[interface] != dict_interface and dict_interface is not None:
+            raise ILDictionaryParseError("Dictionary cannot be used for the chosen communication")
+        firmware_version = device.attrib.get("firmwareVersion")
+        product_code = device.attrib.get("ProductCode")
+        if product_code is not None and product_code.isdecimal():
+            product_code = int(product_code)
+        else:
+            product_code = None
+        part_number = device.attrib.get("PartNumber")
+        revision_number = device.attrib.get("RevisionNumber")
+        if revision_number is not None and revision_number.isdecimal():
+            revision_number = int(revision_number)
+        else:
+            revision_number = None
+        return DictionaryDescriptor(firmware_version, product_code, part_number, revision_number)
+
     def read_dictionary(self) -> None:
         try:
             with open(self.path, "r", encoding="utf-8") as xdf_file:
                 tree = ET.parse(xdf_file)
-        except FileNotFoundError:
-            raise FileNotFoundError(f"There is not any xdf file in the path: {self.path}")
+        except FileNotFoundError as e:
+            raise FileNotFoundError(f"There is not any xdf file in the path: {self.path}") from e
         root = tree.getroot()
 
         device = root.find(self.__DICT_ROOT_DEVICE)